requirements/requirements.txt
Flask-JWT==0.3.2
PyMySQL==0.9.3
flask==1.0.2
flask-restful==0.3.7
flask-sqlalchemy==2.3.2
sqlalchemy-utils==0.33.11
